DDMenu separate component created and removed animation from button
