+++ src/component/Header/NavBarItem/index.tsx
@@ -3,23 +3,7 @@ import { FC } from 'react';
 import Link from 'next/link';
 import styles from './styles.module.css';
 import { useState } from 'react';
-
-interface DDMenuProps {
-    list: string[];
-    offsetX?: number;
-    offsetY?: number;
-}
-const DDMenu: React.FC<DDMenuProps> = ({ list, offsetX = 0, offsetY = 28 }) => {
-    return (
-        <div className={styles.dropdown} style={{ transform: `translateX(${offsetX}px) translateY(${offsetY}px)` }} aria-label='language menu'>
-            {list.map((item, index) => {
-                return (
-                    <Link href="" aria-label={item} className={styles.ddItem} key={index}>{item}</Link>
-                )
-            })}
-        </div>
-    );
-};
+import DDMenu from  '../../common/DropdownMenu/index';
 
 const list = ["ENG", "SPN", "DTC"];
