+++ src/component/common/Button/styles.module.css
@@ -4,7 +4,6 @@
     background-color: var(--text-primary-color);
     border: none;
     color: var(--primary-background-dark-color);
-    transition-duration: background-color 0.4, scale 0.4;
     cursor: pointer;
     padding-left: 23px;
     padding-right: 23px;
@@ -21,5 +20,4 @@
 
 .button:hover {
     background-color: var(--text-button-background-color-secondary);
-    scale: 105%;
 }
