+++ src/component/common/DropdownMenu/styles.module.css
@@ -1,0 +1,18 @@
+@import url("../../../global.module.css");
+.dropdown {
+    position: absolute;
+    display: flex;
+    flex-direction: column;
+    background-color: var(--text-icon-color);
+    align-items: center;
+    border-radius: 15px;
+    min-width: 200px;
+    padding-top: 1rem;
+    padding-bottom: 1rem;
+    z-index: 100;
+}
+
+.ddItem {
+    color: var(--text-tertiary-color);
+    font-weight: 600;
+}
